requirements.txt
aioredis==2.0.1
async-timeout==5.0.1
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.1
redis==5.2.1
rsa==4.9
Telethon==1.38.1
typing_extensions==4.12.2
